Factor out WalletError to own module
