+++ packages/server-wallet/src/handlers/update-channel.ts
@@ -5,7 +5,7 @@ import {ChannelId} from '@statechannels/client-api-schema';
 
 import {SignState, signState} from '../protocols/actions';
 import {ChannelState} from '../protocols/state';
-import {WalletError, Values} from '../errors';
+import {WalletError, Values} from '../errors/wallet-error';
 
 type ChannelStateWithSupported = ChannelState & {
   supported: SignedStateWithHash;
